LFS-79: Servlet for listing users  Add correct dependencies, add bundle to java file, change from resourceTypes to Paths.
