+++ modules/user-viewer/pom.xml
@@ -78,4 +78,35 @@
       </plugin>
     </plugins>
   </build>
+  <dependencies>
+
+  	<dependency>
+  		<groupId>javax.json</groupId>
+  		<artifactId>javax.json-api</artifactId>
+  		<version>1.1.4</version>
+  	</dependency>
+  	<dependency>
+  		<groupId>org.osgi</groupId>
+  		<artifactId>
+  			org.osgi.service.component.annotations
+  		</artifactId>
+  		<version>1.4.0</version>
+  	</dependency>
+  	<dependency>
+  		<groupId>org.apache.sling</groupId>
+  		<artifactId>org.apache.sling.servlets.annotations</artifactId>
+  		<version>1.2.4</version>
+  	</dependency>
+  	<dependency>
+  		<groupId>org.apache.sling</groupId>
+  		<artifactId>org.apache.sling.api</artifactId>
+  		<version>2.3.0</version>
+  	</dependency>
+  	<dependency>
+  		<groupId>javax.servlet</groupId>
+  		<artifactId>javax.servlet-api</artifactId>
+  		<version>3.1.0</version>
+  	</dependency>
+  	
+  </dependencies>
 </project>
